feat: CommentRestController memberId 검증 로직 추가

+++ src/main/java/world/meta/sns/api/controller/CommentRestController.java
@@ -1,16 +1,21 @@
 package world.meta.sns.api.controller;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import world.meta.sns.api.common.mvc.CustomResponse;
 import world.meta.sns.api.security.core.userdetails.PrincipalDetails;
+import world.meta.sns.core.comment.dto.CommentDeleteDto;
 import world.meta.sns.core.comment.dto.CommentRequestDto;
 import world.meta.sns.core.comment.dto.CommentUpdateDto;
 import world.meta.sns.core.comment.service.CommentService;
 
+import static world.meta.sns.api.common.enums.ErrorResponseCodes.MEMBER_RESOURCE_FORBIDDEN;
+
 @RestController
 @RequiredArgsConstructor
+@Slf4j
 public class CommentRestController {
 
     private final CommentService commentService;
@@ -19,10 +24,12 @@ public class CommentRestController {
     public CustomResponse saveComment(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable("boardId") Long boardId, 
                                       @RequestBody CommentRequestDto commentRequestDto) {
 
-        commentRequestDto.setMemberId(principalDetails.getMember().getId());
-        commentRequestDto.setBoardId(boardId);
+        if (!principalDetails.getMember().getId().equals(commentRequestDto.getMemberId())) {
+            log.error("[saveComment] 해당 사용자는 접근 권한이 없습니다.");
+            return new CustomResponse.Builder(MEMBER_RESOURCE_FORBIDDEN).build();
+        }
 
-        commentService.saveComment(commentRequestDto);
+        commentService.saveComment(boardId, commentRequestDto);
 
         return new CustomResponse.Builder().build();
     }
@@ -31,14 +38,24 @@ public class CommentRestController {
     public CustomResponse updateComment(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable("commentId") Long commentId,
                                         @RequestBody CommentUpdateDto commentUpdateDto) {
 
-        commentUpdateDto.setMemberId(principalDetails.getMember().getId());
+        if (!principalDetails.getMember().getId().equals(commentUpdateDto.getMemberId())) {
+            log.error("[updateComment] 해당 사용자는 접근 권한이 없습니다.");
+            return new CustomResponse.Builder(MEMBER_RESOURCE_FORBIDDEN).build();
+        }
+
         commentService.updateComment(commentId, commentUpdateDto);
 
         return new CustomResponse.Builder().build();
     }
 
     @DeleteMapping("/api/v1/comments/{commentId}")
-    public CustomResponse deleteComment(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable("commentId") Long commentId) {
+    public CustomResponse deleteComment(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable("commentId") Long commentId,
+                                        @RequestBody CommentDeleteDto commentDeleteDto) {
+
+        if (!principalDetails.getMember().getId().equals(commentDeleteDto.getMemberId())) {
+            log.error("[deleteComment] 해당 사용자는 접근 권한이 없습니다.");
+            return new CustomResponse.Builder(MEMBER_RESOURCE_FORBIDDEN).build();
+        }
 
         commentService.deleteComment(commentId, principalDetails.getMember().getId());
